pyproject.toml
[project]
name = "colour-science"
version = "0.0.0"
description = "Package for color space and color-matching."
readme = "README.md"
requires-python = ">=3.10"


# Dependencies that will be installed by uv
dependencies = [
    "colorspace>=0.5.0",
    "polars>=1.7.1",
    "colour-science>=0.4.4",
    "matplotlib>=3.9.2",
    "scipy>=1.14.1",
    "plotly>=5.24.1",
    "scikit-learn>=1.5.2",
    "networkx>=3.4",
    "pandas>=2.2.3",
    "pandera>=0.20.4",
    "pyyaml>=6.0.2",
    "black>=24.10.0",
    "frictionless>=5.18.0",
    "seaborn>=0.13.2",
    "jupyter",
    "papermill>=2.6.0",
]

# Optional dependencies
[project.optional-dependencies]
dev = [
    "codespell>=2.3.0",
    "dunamai>=1.22.0",
    "ipdb>=0.13.13",
    "ipython>=8.26.0",
    "jupyter>=1.1.1",
    "mypy>=1.11.2",
    "pre-commit>=3.8.0",
    "ruff>=0.6.2",
    "semver>=3.0.2",
    "typer>=0.12.5",
]
test = [
    "pytest>=6.2.5",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-lazy-fixtures>=1.0.0",
    "coverage>=7.4.0",
]
